# openapi.yaml
openapi: 3.1.0
info:
  title: Industrial Engineering Copilot API
  version: 0.1.0
  description: |
    From Engineering Requirements to Product Decisions.

    Phase 1 exposes health endpoints and foundational schemas.
    Phase 2 exposes read-only product catalog APIs for manufacturers, categories,
    products, specifications, and synthetic pricing.
    Phase 3 exposes RFQ requirement extraction via Groq structured output.
    Phase 4 exposes deterministic engineering validation (PASS/FAIL/UNKNOWN).
    Phase 6 exposes datasheet ingestion, pgvector retrieval, and requirement evidence.
    Phase 7 exposes deterministic quotation generation from catalog pricing.
    Phase 5A exposes deterministic product recommendation and decision support.
  contact:
    name: Industrial Engineering Copilot
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/quotations:
    get:
      tags: [Quotations]
      summary: List quotations
      operationId: listQuotations
      responses:
        "200":
          description: Quotation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationListResponse"
    post:
      tags: [Quotations]
      summary: Create a quotation draft from a validated product
      operationId: createQuotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuotationRequest"
      responses:
        "200":
          description: Quotation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation, pricing, or request error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/quotations/{quotation_id}:
    get:
      tags: [Quotations]
      summary: Get quotation detail
      operationId: getQuotation
      parameters:
        - name: quotation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Quotation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationResponse"
        "404":
          description: Quotation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/quotations/{quotation_id}/status:
    patch:
      tags: [Quotations]
      summary: Update quotation review status
      operationId: updateQuotationStatus
      parameters:
        - name: quotation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuotationStatusRequest"
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationResponse"
        "404":
          description: Quotation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/quotations/{quotation_id}/communication:
    get:
      tags: [Quotations]
      summary: Get quotation customer communication draft
      operationId: getQuotationCommunication
      parameters:
        - name: quotation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Communication draft if present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationCommunicationResponse"
    post:
      tags: [Quotations]
      summary: Create quotation customer communication draft
      operationId: createQuotationCommunication
      parameters:
        - name: quotation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuotationCommunicationRequest"
      responses:
        "200":
          description: Communication draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationCommunicationResponse"
    patch:
      tags: [Quotations]
      summary: Update quotation customer communication draft
      operationId: updateQuotationCommunication
      parameters:
        - name: quotation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuotationCommunicationRequest"
      responses:
        "200":
          description: Updated communication draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationCommunicationResponse"
  /api/v1/quotations/{quotation_id}/send:
    post:
      tags: [Quotations]
      summary: Send quotation to customer after sales review
      operationId: sendQuotationToCustomer
      parameters:
        - name: quotation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Quotation marked sent and follow-up created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendQuotationResponse"
        "422":
          description: Send blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/sales/follow-ups:
    get:
      tags: [Sales]
      summary: List sales follow-ups
      operationId: listSalesFollowUps
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, COMPLETED]
        - name: due_date
          in: query
          schema:
            type: string
            format: date
        - name: sent_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Follow-up list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesFollowUpListResponse"
  /api/v1/sales/follow-ups/{follow_up_id}:
    patch:
      tags: [Sales]
      summary: Update sales follow-up
      operationId: updateSalesFollowUp
      parameters:
        - name: follow_up_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalesFollowUpRequest"
      responses:
        "200":
          description: Updated follow-up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesFollowUpResponse"
components:
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...
    StructuredRequirement:
      type: object
      required: [spec_key, operator]
      properties:
        spec_key:
          type: string
          description: Canonical engineering specification key
        operator:
          type: string
          enum: [eq, neq, gt, gte, lt, lte, contains, in]
        value:
          oneOf:
            - type: boolean
            - type: number
            - type: string
            - type: array
              items:
                type: string
          nullable: true
        unit:
          type: string
          nullable: true
        required:
          type: boolean
          default: true
        priority:
          type: string
          enum: [must_have, preferred]
          nullable: true
        source_text:
          type: string
          nullable: true
...
    ValidationStatus:
      type: string
      enum: [PASS, FAIL, UNKNOWN]
...
    QuotationStatus:
      type: string
      enum: [DRAFT, REVIEWED, APPROVED, READY_TO_SEND, SENT]
    CommunicationStatus:
      type: string
      enum: [DRAFT, READY_TO_SEND, SENT]
    FollowUpPriority:
      type: string
      enum: [P0, P1, P2]
    FollowUpStatus:
      type: string
      enum: [OPEN, COMPLETED]
    CreateQuotationRequest:
      type: object
      required: [customer_name, product_id, quantity, requirements]
      properties:
        customer_name:
          type: string
        customer_email:
          type: string
          nullable: true
        customer_reference:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        discount_percent:
          type: string
          default: "0"
        validity_days:
          type: integer
          default: 30
          minimum: 1
        requirements:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/StructuredRequirement"
    QuotationLineResponse:
      type: object
      required:
        - id
        - product_id
        - model_number
        - description
        - quantity
        - unit_price
        - currency
        - discount_percent
        - line_subtotal
        - line_total
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        model_number:
          type: string
        description:
          type: string
        quantity:
          type: integer
        unit_price:
          type: string
        currency:
          type: string
        discount_percent:
          type: string
        line_subtotal:
          type: string
        line_total:
          type: string
    QuotationResponse:
      type: object
      required:
        - id
        - quotation_number
        - status
        - customer_name
        - currency
        - validity_days
        - lead_time_days
        - technical_status
        - subtotal
        - discount_percent
        - discount_amount
        - total
        - lines
        - validation_snapshot
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        quotation_number:
          type: string
        status:
          $ref: "#/components/schemas/QuotationStatus"
        customer_name:
          type: string
        customer_email:
          type: string
          nullable: true
        customer_reference:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        currency:
          type: string
        validity_days:
          type: integer
        lead_time_days:
          type: integer
        technical_status:
          $ref: "#/components/schemas/ValidationStatus"
        subtotal:
          type: string
        discount_percent:
          type: string
        discount_amount:
          type: string
        total:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/QuotationLineResponse"
        validation_snapshot:
          type: object
          additionalProperties: true
        evidence_snapshot:
          type: object
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    QuotationListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationResponse"
        total:
          type: integer
    UpdateQuotationStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/QuotationStatus"
    QuotationCommunicationResponse:
      type: object
      required:
        - id
        - quotation_id
        - customer_name
        - customer_email
        - subject
        - body
        - status
        - demo_mode
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        quotation_id:
          type: string
          format: uuid
        customer_name:
          type: string
        customer_email:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          $ref: "#/components/schemas/CommunicationStatus"
        sent_at:
          type: string
          format: date-time
          nullable: true
        demo_mode:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateQuotationCommunicationRequest:
      type: object
      properties:
        customer_email:
          type: string
          nullable: true
    UpdateQuotationCommunicationRequest:
      type: object
      properties:
        customer_email:
          type: string
          nullable: true
        subject:
          type: string
        body:
          type: string
        status:
          $ref: "#/components/schemas/CommunicationStatus"
    SendQuotationResponse:
      type: object
      required: [quotation_id, communication, send_detail, demo_mode]
      properties:
        quotation_id:
          type: string
          format: uuid
        communication:
          $ref: "#/components/schemas/QuotationCommunicationResponse"
        send_detail:
          type: string
        demo_mode:
          type: boolean
    FollowUpQuotationSummary:
      type: object
      required: [id, quotation_number, status, currency, total]
      properties:
        id:
          type: string
          format: uuid
        quotation_number:
          type: string
        status:
          type: string
        currency:
          type: string
        total:
          type: string
        sent_at:
          type: string
          format: date-time
          nullable: true
    SalesFollowUpResponse:
      type: object
      required:
        - id
        - quotation_id
        - customer_name
        - customer_email
        - quotation_number
        - follow_up_date
        - priority
        - status
        - quotation
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        quotation_id:
          type: string
          format: uuid
        customer_name:
          type: string
        customer_email:
          type: string
        quotation_number:
          type: string
        follow_up_date:
          type: string
          format: date
        priority:
          $ref: "#/components/schemas/FollowUpPriority"
        status:
          $ref: "#/components/schemas/FollowUpStatus"
        notes:
          type: string
          nullable: true
        quotation:
          $ref: "#/components/schemas/FollowUpQuotationSummary"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SalesFollowUpListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SalesFollowUpResponse"
        total:
          type: integer
    UpdateSalesFollowUpRequest:
      type: object
      properties:
        follow_up_date:
          type: string
          format: date
        priority:
          $ref: "#/components/schemas/FollowUpPriority"
        status:
          $ref: "#/components/schemas/FollowUpStatus"
        notes:
          type: string
          nullable: true